Split Ops out of the package object
